Fit linear models

Fit linear models

Usage

fit_linear_model(data, formula, ..., na.action = NULL)

Arguments

data

an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which lm is called.

formula

an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted. The details of model specification are given under ‘Details’.

...

For lm(): additional arguments to be passed to the low level regression fitting functions (see below).

na.action

a function which indicates what should happen when the data contain NAs. The default is set by the na.action setting of options, and is na.fail if that is unset. The ‘factory-fresh’ default is na.omit. Another possible value is NULL, no action. Value na.exclude can be useful.

Examples

x <- 1:3
y <- 2 * x

fit_linear_model(y ~ x)
#> Error in eval(mf, parent.frame()): argument "formula" is missing, with no default